Move the trackpoints call into its own controller (and rename to tracepoints)
